gcp/requirements.txt
datasets==1.17.0
hydra-core==1.1.1
scikit_learn==1.0.2
setuptools==58.0.4
torch==1.10.0
transformers==4.15.0
google-cloud-storage==2.1.0